Breakup model/config/source_config.py into smaller modules
