--- src/koza/model/__init__.py
@@ -1,3 +1,0 @@
-"""
-Biolink Dataclasses
-"""
