+++ tests/integration/test_validator.py
@@ -12,7 +12,7 @@ import yaml
 import pytest
 
 from koza.cli_utils import transform_source
-from koza.model.config.source_config import OutputFormat
+from koza.model.formats import OutputFormat
 
 # pytest.skip("LinkML issue with `category` slot has `designates_type: true`", allow_module_level=True)
