+++ tests/unit/test_jsonreader.py
@@ -3,7 +3,8 @@ from pathlib import Path
 
 import pytest
 from koza.io.reader.json_reader import JSONReader
-from koza.model.config.source_config import FormatType, JSONReaderConfig
+from koza.model.formats import InputFormat
+from koza.model.reader import JSONReaderConfig
 
 test_zfin_data = Path(__file__).parents[1] / "resources" / "source-files" / "test_BGI_ZFIN.json.gz"
 
@@ -15,7 +16,7 @@ json_path = [
 
 def test_normal_case():
     config = JSONReaderConfig(
-        format=FormatType.json,
+        format=InputFormat.json,
         json_path=json_path,
         files=[],
     )
@@ -27,7 +28,7 @@ def test_normal_case():
 
 def test_required_properties():
     config = JSONReaderConfig(
-        format=FormatType.json,
+        format=InputFormat.json,
         json_path=json_path,
         required_properties=["name", "basicGeneticEntity.primaryId"],
         files=[],
@@ -42,7 +43,7 @@ def test_required_properties():
 
 def test_missing_req_property_raises_exception():
     config = JSONReaderConfig(
-        format=FormatType.json,
+        format=InputFormat.json,
         json_path=json_path,
         required_properties=["fake_prop"],
         files=[],
